r: remove useless module
