+++ core/src/lib.rs
@@ -22,7 +22,6 @@ pub mod databases {
     pub mod table;
     pub mod table_schema;
     pub mod remote_databases {
-        pub mod get_remote_database;
         pub mod remote_database;
         pub mod snowflake;
     }
